requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
openai==1.54.3
python-dotenv==1.0.1
python-multipart==0.0.12
supabase==2.10.0
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.7
